Remove neovim colorscheme
Colorscheme is no longer set in neovim, as it is being set in the
terminal (kitty). When trying out mosh, the colors were all wrong which
makes me this it is [this bug](https://github.com/mobile-shell/mosh/issues/928)
but skipping the colorscheme meant that the colors were set properly.