Move and split neovim configuration
This moves various neovim configuration into a subfolder, and then
splits it into base, lite and full configurations. The key difference is
that lite only includes a smaller subset of treesitter grammars which
will hopefully be faster to install from scratch. For a home machine
which rebuilds the same configuraiton many times, full is nice.