More cleanup, better paths
This removes some not needed configuration, fixes ruby-lsp but also
sets paths so that we use the "home" versions of some packages,
particular node and typescript. This is meant to make the language
servers work, even when the version of node is different (and old)

ruby-lsp seems to want to use the version of ruby that the code base
uses, so that should use the project's versions.