Refactor for options
This is a larger refactor so that all modules use options rather than
just needing to import. This is a bit more explicit but also means that
all modules can hopefully be exported and reused intentionally. The
usage of options means that neovim can go back to being a single file.